Add changes to CoockingInfoSchema and CompostableInfoSchema

+++ lists/Compostable.js
@@ -1,4 +1,4 @@
-const { Relationship, Integer } = require('@keystonejs/fields');
+const { Relationship, Float } = require('@keystonejs/fields');
 
 module.exports = {
   fields: {
@@ -6,6 +6,6 @@ module.exports = {
     tool: { type: Relationship, ref: 'Item.composter' },
     result: { type: Relationship, ref: 'ItemCount' },
     composting: { type: Relationship, ref: 'Item.composting' }, 
-    amountPerStack: { type: Integer },
+    amountPerStack: { type: Float },
   }
 };
